Factor file chooser result callback into separate class
